Initial support for running WPT as a relay server (and integrating with one).
Right now you can submit tests and if a location is configured as a remote relay the test will be forwarded and run on the remote relay.  You can also download and delete tests from a relay server.  

The only remaining work is in the status check, if a test is a remote one, forward the status check to the relay and if the test is complete download the test and delete it from the relay server.