Make test labels clickable when tests are still being run
Test labels are clickable on the comparison page when the results are in, but not clickable while the tests are still being run. Labels should always link to the test results page (or to a particular run, when available).

Need to override the color of the anchor from black to white for legibility.