Fixed #192
Added support for pulling from multiple locations for a given test agent.  This update breaks compatibility with urlblast < 375 so an urlblast update must be pushed along with this update.

On the agent side, just supply a comma-separated list of the test locations that that agent can support tests for:

location=wptdriver,IE_10

for example.  This allows you to have one location that has all of the browsers that can be installed regardless of IE version as well as distinct locations for each version of IE.  It is important that the browser= line in locations.ini matches the browser in wptdriver.ini, even for the IE location.