Changed the profile cleanup to just delete the tabs and cache.
The original code deleted the tabs.  I changed that to nuke the entire profile (minus the shared prefs) which generally worked well but the data reduction proxy settings don't like starting from a completely nuked profile (command-line flag doesn't work and the UI for setting it goes away).

Changing the profile code back to the original logic but also adding the clearing of the cache directory is a compromise that still fully clears the caches and also allows the data reduction proxy settings to work.

The main risk is that since we aren't clearing all of the settings, then whatever the browser settings are configured for manually will apply to all tests using that profile (maybe a good thing but it's an additional check during setup to make sure you don't turn something on).