Close browser between first and repeat view.
Changed the flow for repeat view tests to close the browser between
first and repeat view (but not clear the cache for the repeat view).

This matches the desktop behavior and is preferred because otherwise the
browser's in-memory cache will be used and it has very different caching
behavior.

BUG=234
R=klm@google.com

Review URL: https://codereview.appspot.com/103130044