Added ability to set timeout value on per-test basis
This allows users to set a query parmater "timeout=%d" in runtest.php,
where %d is the amount of time to test should be allowed to run for
before web page test force quits the test.

The results of this test up to this point of timeout will be recorded.
If the test times out, the result code for the step that that timeout
occured on will be 99996.

In order to ensure the security of public-facing WPT instances, a required
parameter "EnablePerTestTimeout" must be added to the wptdriver.ini file.