Added ability to set timeout value on per-test basis
This allows users to set a query parmater "timeout=%d" in runtest.php,
where %d is the amount of time to test should be allowed to run for
before web page test force quits the test.

The results of this test up to this point of timeout will be recorded.
If the test times out, the result code for the step that that timeout
occured on will be 99998.

Manual Testing:
-Verified new code has no effect when EnableUserSetTimeout is not included
in wptdriver.ini file
-Verified test times out when timit limit is reached
-Verified test does not timeout when time limit is not reached
-Verified negative timeout values are disgarded
-Verified that partial test results are returned on multi-step test run