Add Ensighten tag manager as CDN
When testing http://www.nature.com/nplants/ recently I noticed Ensighten
tag manager was not within the list of CDN's.

As far as I am aware Ensighten are distributing content to different
locations and serving from the relevant one.

Slightly out of date but here is a map of Ensightens global network
https://www.ensighten.com/sites/default/files/Ensighten-Tag-Delivery-Network.pdf