Fixed #523 - Fixed timing for multi-step scripts and repeat view tests for Chrome HTTPS data.
The times were baselined against the step start time but that changes for each step and needed to be based against a fixed time.  For repeat view tests, Chrome pre-resolves domains so we need to also sync timestamps for connections (if it starts preconnecting then we may be in trouble).