Adjust page timing based on request start time
Page metrics are based on the performance metrics (navigation start)
whereas request timing are based on the first request start time for a
given page. There can be a substantial difference between those 2
timings. This commit fixes the page timing by computing an offset and
applying it to all page metrics.