Unregister hooks when measurement over
When a measurment times out and a substential number of requests
are still being issued by the browser, the wpt hook
can become unresponsive while it processes all HTTP requests and
hook callbacks making it miss the deadline of 40s. Passed this timeout,
the WPT driver kills the browser and all results are lost.

This PR unregister the wpt hook from all windows hook and try to lesser
the load of the HTTP server by disabling netlog calls.