Exposes screenshots into the HAR
This PR exposes exposes 3 types of screenshots:

 1. Session results: the very last screen for a given measurement. Saved
 in har.log._resultScreenshot
 2. Each page visually complete in har.log.pages.[].screenshots.[]
 3. Screenshots taken by the script with selenium API. If the screenshot
 is saved to a file, the filename (without its extension) becomes the
 screenshot label.

In addition to the image path, the image sha1 and timestamp are saved in
the same image object. The sha1 is computed directly by the agent when
the image is saved.

This PR also removes some images not used by AppDynamics synthetic
service.