Don't include visually complete when page number mismatch
When doing measurement with Chrome, it is possible that the page number
included in the visually complete image filename exceeds the number of
pages in the HAR due to the page separation bug we have. This PR ignores
all visually complete images if the number of pages are not the same
between devtools and visually complete filename.