result processing: Regard multi-step video frames
If video frames of a multi-step execution are posted to the server,
make sure they are moved to the correct directory.

Before, all videos were moved to video_<run><Cached_?>.
Now, the video folder may get a suffix for the specific step it
corresponds to (e.g. `video_1_Cached_3`.
For the first step, the step suffix is ommited, so file structure
and server versions stay compatible.

A `TestFileNames` class was introduced to take the responsibilty
for creating file names and paths for test-related files.
This way, the logic of file names and the file names itself
get encapsulated and can get tested.