xmlResult: Added unit test for median output
The encapsuled logic finally enables to test at least
the generation of the run and median outputs in XML
gemeration.

For a proper test I exported data and corresponding XML
export from real results and used them in a unit test.

The test can now be used to verify that for given the result
is compatible.

Compatible means in this context that the generated XML should
include at least all nodes that are expected, i.e. were generated
before.