New setting: EC2.IdleTerminateMinutes
This setting is for Linux EC2 instances, which are now billed by the second,
with a minimum of 60 seconds.  The name says it all, it is an int that provides
the number of minutes to let an instance sit idle before terminating it.

I have included comments around this setting, because some caution is needed.
If you push the number too low the agent will not have time to spin up and
ask for tests before being considered idle.

Windows instances are still billed per hour.  When EC2.IdleTerminateMinutes
is not set then the existing termination logic is still applied.