Add support for EC2 private IP addresses
Add an if block to allow the ec2_use_server_private_ip setting to choose whether we want to send the public EIP or the private IP address on the WPT server in the user data for agents so an entire WPT setup could be in a completely private monitoring network. Set host variable to empty at beginning of function to make sure we are always getting/setting the right value either based on settings or on the EC2 meta-data. This should fix a weird bug I saw where $host was somehow already set with a public IP even though the GetSetting('host') didn't have a value and the value couldn't have been otherwise cached.