feat(composer): Add composer file for phpunit
Composer is a package manager for php (similar to npm for JavaScript or
bundler for Ruby, etc.). We can use it to make sure we install the
appropriate dependencies for our project. In this case, I'd like to
install phpunit.

Since this is a developer dependency, this will not break deployments as
they stand, as these files are not required for the application to run,
just for the tests to run