feat: monthly rate limiter
This here monthly rate limiter class is a quick and dirty way to make
sure we don't get hit by more than 50 tests from a specific IP address
of an anonymous user in a 31 day period. We can set those numbers to
higher numbers or lower numbers, though.

This should probably be done with something sturdier (see: redis), but for
right now uses the APCU plugin, which is what we use for our rate
limiting for 3 minute long periods already. It will lose track of things
when the machine gets restarted but I don't think we care _that_ much,
for now.