feat(User): create User class
This is a User model. It has some attributes on it that are important to
understand about the current user as we build web pages, such as their
email, if they are an admin, and whether or not they are a paid user.

Have added this user as $current_user to the exported common.inc file so
it will be referenced in any page that is used