fix(login): product login redirects incorrectly
$_SERVER["HTTP_REFERER"] will include the host but not the entire path for the
"referer"ing website. This means that when somebody tries to log in from said
page, it will take them to a broken page when they're done.

This is not ideal.

Just go back to www site when you log in. The webflow-based product page
doesn't know how to deal with logged in users yet, anyway.