This change removes a max-height on overflow tables, which will mean that column headers will no longer be able to stick as you scroll.
We'd debated having sticky column headers and decided it's better than not in many cases, but the downsides are tables can have overflow that's not immediately obvious. For now, we can disable this for the vertical heights. Horizontal overflow will still have sticky row headers, which is more crucial on mobile sizes.