feat(login): styles
This styles the login page and adds a layer of error notifications. The
error notifications need to be fleshed out more, but this is fine for
now. As for the login, it has some more touches that can be applied, but
this is so other people can also touch it up