chore(refactor): Use types a little stronger
OK, so it's not perfect, but the IDE help is good and I could use the
guidance when digging into something new.

toArray should probably be implemented via a trait. I'll get there. This
is a way to catch a few errors here and there and leave a little less to
humans.

Noting: Handlers doesn't implement the PSR, yet, but this is a way to
move in that direction