feat(history): enabled for logged in user
This change allows our logged in users to take advantage of the same
page that our current logged out users use, except without the indexdb!

We had to change the order in which a few things were loaded, as well,
just so we could make sure the $owner global was set properly. We also
changed types to better match what CP was providing us