fix(account): style plan selector
Please forgive me for the CSS crimes I have committed today.

First: Those selectors. You see. We're replicating UX that would
normally be done with JavaScript and we're doing it _without javascript_
so I feel like this is something that can be overlooked.

Second: We should find a fun and clever way to make this something a
little more reusable because we have this pattern a few times