fix(signup): display error when user signs up
This covers a few different kind of error. If the user manages to put in
incorrect or bad data, this will show. Also, if the user uses an email
that's already been used, an error will show.

This leads us to an interesting place where we need to think out how we
want to handle global (banner) vs local (in the form itself?) errors and
how we make those easy enough to get into