feat(api): delete multiple keys together
This allows us to select multiple delete keys at the same time to delete
together. It uses semantic html in which to do so. This _does_ mean that
the page jumps to the top when you delete. Not super ideal, but it is
what it is.

On top of this, is some logic around deleting things and how checkboxes
should work and how a solid select all situation should work.