feat(signup): smooth out error experience
This allows for an error to pop up at the top of the page when one is
called for during the signup experience. It also fills the information
back in when somebody has an error so they don't have to re-enter it.
It also disables the signup button while submission is happening so
people don't think there's something left to do.