fix(testlog): make test history more resilient
This is a test. We're trying to figure out why test history keeps going
down bigger queries or queries in general.

This involves removing the time limit on the page for logged in users
while simultaneously setting a large timeout for the backend calls.
We're also making the query and things around it a bit more resilient