fix(account): WPT Enterprise
If a user is WPT Enterprise, they do not have a braintree id and thus
get access denied when querying for braintree transaction history. This
is bad because our graphql client blows up on ANY errors existing.

So, let's make a different query if somebody is wpt enterprise to not
query for that if they are