More upgrade plan styling
(THIS CAME WITH SOME SEVERE REBASE FIXES. IT'S NOT PRETTY. It's a lot of
the future. I know. What have we done?)

It's just... the rebase was bad. It was really bad. Sometimes you make
sacrifices. This was one of those days. We lost a lot of our history.
Future generations will cry.