custom metric for script and link elements that could potentially block render.
This is not a guarantee, as these may have been added dynamically by the time this runs. Still, we can use this list to compare in browsers that don't send a render-blocking flag on requests, so we can be more confident that a request that arrived before start render was actually render-blocking. Browsers such as chromium currently give us that information directly so we do not need this in those browsers.