chore(account handler): refactor methods
In this case, we're making it so handlers tend to return something. We
also are breaking validation out into its own method.

Also including attributes, though they are commented out, so we can use
them in the future to designate routes and types of methods better. As
this provides us with some fun metadata we can use.

This leads to easier reading through the code, finding where things are
done, and testing