fix(account): pulse animation
The icon being used within a button where there was a pulse animation
was overwriting the pulse. Given the utility class nature of the pulse,
we're going to use !important.

This is css crimes. I'm aware of the crimes. It's less css felony and
more css misdemeanor, but it is still crimes