fix(account): price could be misleading
Listen, sometimes your user buys a thing and the price says one thing
but they paid a different amount because of taxes. You want to message
that to a user in a way that really lets them know you care.

But at the same time, the service you use requires an extra query that
is very intense to show the applicable taxes and yeesh.

No code is better than no code, but in this case, what if it's just some
text that makes it so logic and network calls don't need to be made?