fix(account): account info not updating
Active contact in user details is cached for a minute and does not
invalidate on updates.

This means that if you change things on the account page and are relying
on that, the caching will bite you.

Adding an additional call to the account page for user info fixes this
problem from our end