feat(api): add header-based api key
Users frequently log their urls that they request. It's a thing that
happens. They do this with headers, less. This allows those that do this
to have logs that don't have api keys littered throughout.

Usage - pass the request header of `X-WPT-API-KEY` (well, any case
insensitive version of this, honestly, in order to follow the spec for
how request header should work) with the value of your api key instead
of passing the `k` query param